Added fix to skip backup and restore instances for TSN nodes
Closes-Bug: #1552174

+++ fabfile/tasks/backup_restore.py
@@ -177,7 +177,7 @@ def backup_cassandra(db_datas, store_db='local', cassandra_backup='full'):
             msg = "Processing the Cassandra DB Backup and puting DB dabckup data into remote host:({HOST}) and default path is ~/contrail_bkup_data/hostname/data/ or backup path is defined  as per testbed.py file  \n "
             remote_host = env.roledefs['backup'][0]
             print (msg.format(HOST=remote_host))
-        OS= sudo("cat /etc/issue  | awk 'FNR == 1 {print $1 }'")
+        OS= sudo("cat /etc/os-release  | awk 'FNR == 1 {print $1 }'| awk  -F '\"' '{print $2}'")
         if OS == 'Ubuntu' :
            snap_path = sudo(
                            "grep  -A 1  'data_file_directories:' --after-context=1  /etc/cassandra/cassandra.yaml")
@@ -260,7 +260,9 @@ def backup_mysql(db_datas, store_db='local'):
             remote_dir_name = final_dir + '%s/' % (host_name)
     with settings(host_string=host):
         mysql_passwd = sudo('cat /etc/contrail/mysql.token')
-        mysql_cmd = 'mysqldump  -u root --password=%s --all-databases  > openstack.sql' % (
+        if exists('openstack.sql') :
+            sudo('rm -r openstack.sql')
+        mysql_cmd = 'mysqldump  -u root --password=%s --all-databases  -–skip-add-locks --ignore-table=mysql.event > openstack.sql' % (
             mysql_passwd)
         sudo(mysql_cmd)
         if store_db == 'local':
@@ -378,45 +380,46 @@ def backup_instances(db_datas, store_db='local'):
     """Backup instances data to all compute nodes """
     host = env.host_string
     global backup_path, final_dir
-    msg = "Processing instances backup and default path for backup data is ~/contrail_bkup_data/hostname/instances  in ({HOST}) \n"
-    with settings(host_string=host):
-        host_name = sudo('hostname')
-        if store_db == 'local':
-            print (msg.format(HOST=host_name))
-        if store_db == 'remote':
-            msg = "Processing instances data Backup and puting instances backup data into remote host:({HOST}) and default path is ~/contrail_bkup_data/hostname/instances or backup path is defined as per testbed.py file \n"
-            remote_host = env.roledefs['backup'][0]
-            print (msg.format(HOST=remote_host))
-        insta_path = sudo("grep  '^state_path'   /etc/nova/nova.conf")
-        instances_path = insta_path.split('=')
-        instances_path = instances_path[-1]
-        instances_path = instances_path.strip()
-        db_path = instances_path
-        execute(verify_disk_space, db_datas, db_path, store_db)
-        remote_instances_path = '%s%s/' % (backup_path, host_name)
-        if db_datas:
-            remote_instances_path = final_dir + '%s/' % (host_name)
-        if store_db == 'local':
-           sudo(
-                'rsync -az --progress   %s/instances    %s' %
-                (instances_path, remote_instances_path))
-        execute(
-            backup_info_file,
-            remote_instances_path,
-            backup_type='Instances')
-        if store_db == 'local':
-            sudo('cp backup_info.txt %s' % remote_instances_path)
-        if store_db == 'remote':
-            remote_path = '%s' % (remote_instances_path)
-            source_path = '%s/instances' % (instances_path)
-            execute(ssh_key_gen)
-            remote_cmd='rsync -avz -e "ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null" --progress    %s   %s:%s' %(  source_path,remote_host,remote_path)
-            sudo(remote_cmd)
-            remote_bk_cmd = 'scp -o "StrictHostKeyChecking no" -r  backup_info.txt  %s:%s' % (
-                remote_host,
-                remote_path)
-            sudo(remote_bk_cmd)
-        sudo('rm -rf backup_info.txt')
+    if host not in env.roledefs['tsn'] :
+        msg = "Processing instances backup and default path for backup data is ~/contrail_bkup_data/hostname/instances  in ({HOST}) \n"
+        with settings(host_string=host):
+            host_name = sudo('hostname')
+            if store_db == 'local':
+                print (msg.format(HOST=host_name))
+            if store_db == 'remote':
+                msg = "Processing instances data Backup and puting instances backup data into remote host:({HOST}) and default path is ~/contrail_bkup_data/hostname/instances or backup path is defined as per testbed.py file \n"
+                remote_host = env.roledefs['backup'][0]
+                print (msg.format(HOST=remote_host))
+            insta_path = sudo("grep  '^state_path'   /etc/nova/nova.conf")
+            instances_path = insta_path.split('=')
+            instances_path = instances_path[-1]
+            instances_path = instances_path.strip()
+            db_path = instances_path
+            execute(verify_disk_space, db_datas, db_path, store_db)
+            remote_instances_path = '%s%s/' % (backup_path, host_name)
+            if db_datas:
+                remote_instances_path = final_dir + '%s/' % (host_name)
+            if store_db == 'local':
+               sudo(
+                    'rsync -az --progress   %s/instances    %s' %
+                    (instances_path, remote_instances_path))
+            execute(
+                backup_info_file,
+                remote_instances_path,
+                backup_type='Instances')
+            if store_db == 'local':
+                sudo('cp backup_info.txt %s' % remote_instances_path)
+            if store_db == 'remote':
+                remote_path = '%s' % (remote_instances_path)
+                source_path = '%s/instances' % (instances_path)
+                execute(ssh_key_gen)
+                remote_cmd='rsync -avz -e "ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null" --progress    %s   %s:%s' %(  source_path,remote_host,remote_path)
+                sudo(remote_cmd)
+                remote_bk_cmd = 'scp -o "StrictHostKeyChecking no" -r  backup_info.txt  %s:%s' % (
+                    remote_host,
+                    remote_path)
+                sudo(remote_bk_cmd)
+            sudo('rm -rf backup_info.txt')
 # end backup_instances
 
 def verify_disk_space(db_datas, db_path, store_db):
@@ -740,7 +743,7 @@ def restore_cassandra(backup_data_path='', store_db='local',cassandra_backup='fu
             snapshot_dir = '~/contrail-data/data/'
         snapshot_list = snapshot_list.split('\r\n')
         snapshot_list = snapshot_list[0]
-        OS= sudo("cat /etc/issue  | awk 'FNR == 1 {print $1 }'")
+        OS= sudo("cat /etc/os-release  | awk 'FNR == 1 {print $1 }'| awk  -F '\"' '{print $2}'")
         if OS == 'Ubuntu' :
            snap_path = sudo(
                            "grep  -A 1  'data_file_directories:' --after-context=1  /etc/cassandra/cassandra.yaml")
@@ -948,51 +951,52 @@ def restore_instances(backup_data_path, store_db='local'):
     """Restore instances data to all compute nodes  """
     global backup_path
     host = env.host_string
-    msg = "Restoring backed-up instances data in ({HOST}).\n"
-    with settings(host_string=host):
-        host_name = sudo('hostname')
-        insta_path = sudo("grep  '^state_path'   /etc/nova/nova.conf")
-    instances_path = insta_path.split('=')
-    instances_path = instances_path[-1]
-    instances_path = instances_path.strip()
-    remote_instances_dir_path = '%s%s/' % (backup_path, host_name)
-    if backup_data_path:
-        if store_db == 'local':
-            with settings(host_string=host):
-                for bk_path in backup_data_path:
-                    snapshot_path = bk_path + '%s/' % (host_name)
-                    if exists('%sinstances/' % snapshot_path):
-                        remote_instances_dir_path = snapshot_path
-        if store_db == 'remote':
-            with settings(host_string=env.roledefs['backup'][0]):
-                for bk_path in backup_data_path:
-                    snapshot_path = bk_path + '%s/' % (host_name)
-                    if exists('%sinstances/' % snapshot_path):
-                        remote_instances_dir_path = snapshot_path
-    if store_db == 'remote':
-        msg = "Restoring instances backed-up  data and getting  backup data from remote host:({HOST}) \n"
-        remote_host = env.roledefs['backup'][0]
-        print (msg.format(HOST=remote_host))
-        remote_path = '%sinstances' % (remote_instances_dir_path)
-        with settings(host_string=env.roledefs['backup'][0]):
-            if not exists(remote_path):
-                print "Remote path doesnot exist ... So aborting the restore nova instances data."
-                execute(start_nova_openstack_compute)
-                raise SystemExit()
-        execute(ssh_key_gen)
+    if host not in env.roledefs['tsn'] :
+        msg = "Restoring backed-up instances data in ({HOST}).\n"
         with settings(host_string=host):
-            remote_cmd='rsync -avz -e "ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null" --progress   %s:%s    %s/' %(  remote_host,remote_path,instances_path)
-            sudo(remote_cmd)
-    with settings(host_string=host):
-        if store_db == 'local':
-            print (msg.format(HOST=host_name))
+            host_name = sudo('hostname')
+            insta_path = sudo("grep  '^state_path'   /etc/nova/nova.conf")
+        instances_path = insta_path.split('=')
+        instances_path = instances_path[-1]
+        instances_path = instances_path.strip()
+        remote_instances_dir_path = '%s%s/' % (backup_path, host_name)
+        if backup_data_path:
+            if store_db == 'local':
+                with settings(host_string=host):
+                    for bk_path in backup_data_path:
+                        snapshot_path = bk_path + '%s/' % (host_name)
+                        if exists('%sinstances/' % snapshot_path):
+                            remote_instances_dir_path = snapshot_path
+            if store_db == 'remote':
+                with settings(host_string=env.roledefs['backup'][0]):
+                    for bk_path in backup_data_path:
+                        snapshot_path = bk_path + '%s/' % (host_name)
+                        if exists('%sinstances/' % snapshot_path):
+                            remote_instances_dir_path = snapshot_path
+        if store_db == 'remote':
+            msg = "Restoring instances backed-up  data and getting  backup data from remote host:({HOST}) \n"
+            remote_host = env.roledefs['backup'][0]
+            print (msg.format(HOST=remote_host))
             remote_path = '%sinstances' % (remote_instances_dir_path)
-            if not exists(remote_path):
-                print "Remote path doesnot exist ... So aborting the restore nova instaces data ."
-                execute(start_nova_openstack_compute)
-                raise SystemExit()
-            sudo('rsync -az --progress  %sinstances  %s/ ' %
-                (remote_instances_dir_path,instances_path ))
+            with settings(host_string=env.roledefs['backup'][0]):
+                if not exists(remote_path):
+                    print "Remote path doesnot exist ... So aborting the restore nova instances data."
+                    execute(start_nova_openstack_compute)
+                    raise SystemExit()
+            execute(ssh_key_gen)
+            with settings(host_string=host):
+                remote_cmd='rsync -avz -e "ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null" --progress   %s:%s    %s/' %(  remote_host,remote_path,instances_path)
+                sudo(remote_cmd)
+        with settings(host_string=host):
+            if store_db == 'local':
+                print (msg.format(HOST=host_name))
+                remote_path = '%sinstances' % (remote_instances_dir_path)
+                if not exists(remote_path):
+                    print "Remote path doesnot exist ... So aborting the restore nova instaces data ."
+                    execute(start_nova_openstack_compute)
+                    raise SystemExit()
+                sudo('rsync -az --progress  %sinstances  %s/ ' %
+                    (remote_instances_dir_path,instances_path ))
 
   # end restore_instances
